Reworking some update logic to prepare for bigger changes.
This update removes some signal/slot communications between the log
database and display widgets in favor of simple timers.  This is to
allow messages to be processed immediately when they are added to the
database instead of being queued and then added in bulk.  That was
mostly done to limit the update rates of the widgets, so it makes
sense that the widgets (or models in this case) take care of the
limiting themselves.