More memory improvements by sharing log message data.
This is another pass at reducing memory usage by sharing log message
data.  We extract numbers out of log messages and store them
seperately, The numbers in the log are replaced with special tokens to
get a generic log message.  The generic message is stored in a shared
cache in the database.  Combined with the previous improvements, I'm
seeing significantly less memory usage (approximately 1/3 as much
memory compared to current swri-robotics master).